fix dt updating, screen jitter bug, add circle2D methods

+++ src/main/java/util/Time.java
@@ -1,10 +1,10 @@
 package util;
 
 public class Time {
-    public static double timeStarted =System.nanoTime();
-    public static double timePassed = 0;
+    public static long timeStarted = System.nanoTime(); // big number
+    public static long timePassed = 0;
 
-    public static double getTime() {
-        return ((System.nanoTime() - timeStarted) * 1E-9);
+    public static float getTime() {
+        return (float) ((System.nanoTime() - timeStarted) * 1E-9);
     }
 }
